propuesta code refactor
